Moves monitor to it's own folder
